According to the cuda documenation (https://docs.nvidia.com/cuda/cuda-math-api/group__CUDA__MATH__DOUBLE.html#group__CUDA__MATH__DOUBLE_1g6d36757715384dc18e0483aa1f04f6c7) pow() gets two double arguments. With the newest VisualStudio Version 16.8 these arguments needs to be passed as double explicitly.

+++ src/gpu/VirtualFluids_GPU/GPU/Cumulant27chim.cu
@@ -255,9 +255,9 @@ extern "C" __global__ void Cumulant_One_preconditioned_errorDiffusion_chim_Comp_
 				(mfbbc - mfbba)) / rho;
 			////////////////////////////////////////////////////////////////////////////////////
 			//the force be with you
-			real fx = forces[0] / (pow(c2o1, level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
-			real fy = forces[1] / (pow(c2o1, level)); //zero;
-			real fz = forces[2] / (pow(c2o1, level)); //zero;
+			real fx = forces[0] / (pow((double)c2o1, (double)level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
+			real fy = forces[1] / (pow((double)c2o1, (double)level)); //zero;
+			real fz = forces[2] / (pow((double)c2o1, (double)level)); //zero;
 			vvx += fx*c1o2;
 			vvy += fy*c1o2;
 			vvz += fz*c1o2;
@@ -1170,9 +1170,9 @@ extern "C" __global__ void Cumulant_One_preconditioned_chim_Comp_SP_27(
 				(mfbbc - mfbba)) / rho;
 			////////////////////////////////////////////////////////////////////////////////////
 			//the force be with you
-			real fx = forces[0] / (pow(c2o1, level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
-			real fy = forces[1] / (pow(c2o1, level)); //zero;
-			real fz = forces[2] / (pow(c2o1, level)); //zero;
+			real fx = forces[0] / (pow((double)c2o1, (double)level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
+			real fy = forces[1] / (pow((double)c2o1, (double)level)); //zero;
+			real fz = forces[2] / (pow((double)c2o1, (double)level)); //zero;
 			vvx += fx*c1o2;
 			vvy += fy*c1o2;
 			vvz += fz*c1o2;
@@ -1943,9 +1943,9 @@ extern "C" __global__ void Cumulant_One_chim_Comp_SP_27(
 				(mfbbc - mfbba)) * OOrho;
 			////////////////////////////////////////////////////////////////////////////////////
 			//the force be with you
-			real fx = forces[0] / (pow(c2o1, level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
-			real fy = forces[1] / (pow(c2o1, level)); //zero;
-			real fz = forces[2] / (pow(c2o1, level)); //zero;
+			real fx = forces[0] / (pow((double)c2o1, (double)level)); //zero;//0.0032653/(pow(two,level)); //0.000000005;//(two/1600000.0) / 120.0; //
+			real fy = forces[1] / (pow((double)c2o1, (double)level)); //zero;
+			real fz = forces[2] / (pow((double)c2o1, (double)level)); //zero;
 			vvx += fx*c1o2;
 			vvy += fy*c1o2;
 			vvz += fz*c1o2;
